// resources/js/Core/Types/ExecutiveOrders/ExecutiveOrders.ts
export interface ExecutiveOrderFormData {
    number: string;
    title: string;
    dateIssued: string;
    description: string;
    category: string;
}

export interface ExecutiveOrder {
    number: string;
    title: string;
    dateIssued: string;
    description: string;
    category: string;
}